Customer Churn Predicition/deployment/requirements.txt
streamlit 
pandas  
seaborn 
matplotlib
Pillow
numpy
tensorflow
feature_engine
scikit-learn==1.2.2